+++ app/src/main/java/com/codemate/koffeemate/ui/userselector/adapter/UserQuickDialAdapter.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.codemate.koffeemate.ui.userselector
+package com.codemate.koffeemate.ui.userselector.adapter
 
 import android.animation.ObjectAnimator
 import android.view.animation.AccelerateDecelerateInterpolator
